테스트 서버와 실제 서버 환경 맞추기 머지

See merge request s09-final/S09P31D209!41

+++ tingting/src/main/java/com/alsif/tingting/concert/service/ConcertService.java
@@ -157,7 +157,7 @@ public class ConcertService {
 	/*
 		콘서트 상세 정보 불러오기
 	 */
-	public ConcertDetailResponseDto findConcertDetail(Long concertSeq, Long userSeq) {
+	public ConcertDetailResponseDto findConcertDetail(Integer concertSeq, Integer userSeq) {
 
 		ConcertDetailResponseDto concertDetailResponseDto
 			= concertRepository.findByConcertDetailsByConcertSeq(concertSeq)
